+++ lib/middle_end/ir_storage.c
@@ -23,7 +23,7 @@ void ir_storage_reset()
 
 void ir_storage_push(const char *name, int32_t ir_idx)
 {
-    hashmap_put(&storage, crc32_string(name), (size_t) ir_idx);
+    hashmap_put(&storage, crc32_string(name), (uint64_t) ir_idx);
 }
 
 int32_t ir_storage_get(const char *name)
